+++ dev/scripts/py/prepend_dev_docs_test_md.py
@@ -1,0 +1,10 @@
+filename = "dev/docs/tests.md"
+
+heading = """<h1 align="center" style="font-weight: bold">
+    Tests
+</h1>\n\n"""
+
+with open(filename, "r+", encoding="utf-8") as f:
+    original = f.read()
+    f.seek(0)
+    f.write(heading + original)